DOI-187 Remove Deprecated code
* Add v2 to jersey application path
* Delete create and update from DAO
* Moving away from versioning API URIs
