+++ disa-api-core/src/main/java/mz/org/fgh/disaapi/core/result/model/LabResult.java
@@ -15,27 +15,15 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
 import javax.persistence.Table;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import mz.co.msaude.boot.frameworks.model.GenericEntity;
-import mz.org.fgh.disaapi.core.result.dao.LabResultDAO;
 
 /**
  * @author Stélio Moiane
  * @author Hélio Machabane
  *
  */
-@NamedQueries({
-		@NamedQuery(name = LabResultDAO.QUERY_NAME.findByLocationCodeAndStatus, query = LabResultDAO.QUERY.findByLocationCodeAndStatus),
-		@NamedQuery(name = LabResultDAO.QUERY_NAME.findByLocationCodeAndStatusSimple, query = LabResultDAO.QUERY.findByLocationCodeAndStatusSimple),
-		@NamedQuery(name = LabResultDAO.QUERY_NAME.findByNid, query = LabResultDAO.QUERY.findByNid),
-		@NamedQuery(name = LabResultDAO.QUERY_NAME.findByRequestId, query = LabResultDAO.QUERY.findByRequestId),
-		@NamedQuery(name = LabResultDAO.QUERY_NAME.findByStatusAndDates, query = LabResultDAO.QUERY.findByStatusAndDates),
-		@NamedQuery(name = LabResultDAO.QUERY_NAME.findByLocationCodeStatusAndNotProcessingCause, query = LabResultDAO.QUERY.findByLocationCodeStatusAndNotProcessingCause) })
 @Entity
 @Table(name = "VlData")
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
@@ -496,23 +484,5 @@ public abstract class LabResult extends GenericEntity {
 
 	public void setRegisteredDateTime(LocalDateTime registeredDateTime) {
 		this.registeredDateTime = registeredDateTime;
-	}
-
-	/**
-	 * @deprecated Added in 2.1.3 to support backward compatibility.
-	 */
-	@Deprecated
-	@JsonProperty("viralLoadStatus")
-	public LabResultStatus getViralLoadStatus() {
-		return getLabResultStatus();
-	}
-
-	/**
-	 * @deprecated Added in 2.1.3 to support backward compatibility.
-	 */
-	@Deprecated
-	@JsonProperty("viralLoadResultDate")
-	public LocalDateTime getViralLoadResultDate() {
-		return getLabResultDate();
 	}
 }
