+++ disa-api-core/src/test/java/mz/co/fgh/disaapi/core/fixturefactory/ViralLoadTemplate.java
@@ -12,7 +12,6 @@ import br.com.six2six.fixturefactory.Rule;
 import br.com.six2six.fixturefactory.loader.TemplateLoader;
 import mz.co.msaude.boot.frameworks.model.EntityStatus;
 import mz.org.fgh.disaapi.core.result.model.HIVVLLabResult;
-import mz.org.fgh.disaapi.core.result.model.LabResult;
 import mz.org.fgh.disaapi.core.result.model.LabResultStatus;
 import mz.org.fgh.disaapi.core.result.model.TypeOfResult;
