+++ disa-api-core/src/test/java/mz/co/fgh/disaapi/core/viralload/integ/test/unit/viralload/ViralLoadServiceUnitTest.java
@@ -2,10 +2,7 @@ package mz.co.fgh.disaapi.core.viralload.integ.test.unit.viralload;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import javax.inject.Inject;
@@ -23,7 +20,6 @@ import mz.co.fgh.disaapi.core.fixturefactory.ViralLoadTemplate;
 import mz.co.msaude.boot.frameworks.exception.BusinessException;
 import mz.co.msaude.boot.frameworks.fixturefactory.EntityFactory;
 import mz.co.msaude.boot.frameworks.model.EntityStatus;
-import mz.org.fgh.disaapi.core.exception.NotFoundBusinessException;
 import mz.org.fgh.disaapi.core.result.model.HIVVLLabResult;
 import mz.org.fgh.disaapi.core.result.model.LabResult;
 import mz.org.fgh.disaapi.core.result.model.LabResultStatus;
@@ -41,7 +37,7 @@ public class ViralLoadServiceUnitTest extends AbstractUnitServiceTest {
 
     @BeforeClass
     public static void setUp() {
-            FixtureFactoryLoader.loadTemplates("mz.co.fgh.disaapi.core.fixturefactory");
+        FixtureFactoryLoader.loadTemplates("mz.co.fgh.disaapi.core.fixturefactory");
     }
 
     @Test
@@ -82,60 +78,6 @@ public class ViralLoadServiceUnitTest extends AbstractUnitServiceTest {
                 propertyValues);
 
         assertThat(viralLoad.getReasonForTest()).isEqualTo("Routine");
-
-    }
-
-    @Test(expected = BusinessException.class)
-    public void updateLabResultShouldFailWhenViralLoadProcessed() throws BusinessException {
-
-        LabResult viralLoad = EntityFactory.gimme(HIVVLLabResult.class, ViralLoadTemplate.PROCESSED);
-
-        List<String> requestIds = Arrays.asList(viralLoad.getRequestId());
-        Mockito.when(viralLoadRepository.findByRequestIdInAndEntityStatus(requestIds, EntityStatus.ACTIVE))
-                .thenReturn(Arrays.asList(viralLoad));
-
-        Map<String, Object> propertyValues = new HashMap<>();
-        propertyValues.put("labResultStatus", "PENDING");
-        propertyValues.put("healthFacilityLabCode", "1040107");
-        propertyValues.put("requestingFacilityName", "CS 24 de Julho");
-        this.viralLoadService.updateLabResult(
-                viralLoad,
-                propertyValues);
-
-    }
-
-    @Test(expected = BusinessException.class)
-    public void updateLabResultShouldFailWhenPropertyEmpty() throws BusinessException {
-
-        LabResult viralLoad = EntityFactory.gimme(HIVVLLabResult.class, ViralLoadTemplate.NOT_PROCESSED);
-
-        List<String> requestIds = Arrays.asList(viralLoad.getRequestId());
-        Mockito.when(viralLoadRepository.findByRequestIdInAndEntityStatus(requestIds, EntityStatus.ACTIVE))
-                .thenReturn(Arrays.asList(viralLoad));
-
-        Map<String, Object> propertyValues = new HashMap<>();
-        propertyValues.put("labResultStatus", "");
-        this.viralLoadService.updateLabResult(
-                viralLoad,
-                propertyValues);
-    }
-
-    @Test(expected = NotFoundBusinessException.class)
-    public void updateLabResultShouldFailWhenViralNotFound() throws BusinessException {
-
-        LabResult viralLoad = EntityFactory.gimme(HIVVLLabResult.class, ViralLoadTemplate.PROCESSED);
-
-        List<String> requestIds = Arrays.asList(viralLoad.getRequestId());
-        Mockito.when(viralLoadRepository.findByRequestIdInAndEntityStatus(requestIds, EntityStatus.ACTIVE))
-                .thenReturn(Collections.emptyList());
-
-        Map<String, Object> propertyValues = new HashMap<>();
-        propertyValues.put("labResultStatus", "PENDING");
-        propertyValues.put("healthFacilityLabCode", "1040107");
-        propertyValues.put("requestingFacilityName", "CS 24 de Julho");
-        this.viralLoadService.updateLabResult(
-                viralLoad,
-                propertyValues);
 
     }
 }
