+++ disa-api-integ/src/main/java/mz/org/fgh/disaapi/integ/resources/config/JerseyConfig.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Service;
 
 import mz.org.fgh.disaapi.integ.resources.converter.LocalDateTimeParamConverterProvider;
 import mz.org.fgh.disaapi.integ.resources.orgunit.OrgUnitResource;
-import mz.org.fgh.disaapi.integ.resources.viralload.ViralLoadResource;
-import mz.org.fgh.disaapi.integ.resources.viralload.ViralLoadResourceV2;
+import mz.org.fgh.disaapi.integ.resources.orgunit.OrgUnitResourceV2;
 import mz.org.fgh.disaapi.integ.resources.viralload.LabResultResource;
+import mz.org.fgh.disaapi.integ.resources.viralload.ViralLoadResource;
 
 /**
  * @author Stélio Moiane
@@ -23,7 +23,7 @@ public class JerseyConfig extends ResourceConfig {
 
 	public JerseyConfig() {
 		this.register(ViralLoadResource.class);
-		this.register(ViralLoadResourceV2.class);
+		this.register(OrgUnitResourceV2.class);
 		this.register(LabResultResource.class);
 		this.register(OrgUnitResource.class);
 		this.register(LocalDateTimeParamConverterProvider.class);
